pyproject.toml
[project]
name = "javis"
version = "0.1.0"
description = "Jarvis AI Framework - A modular AI assistant framework with multi-language support"
authors = [
    {name = "JHM",email = "jhm@example.com"}
]
license = {text = "MIT"}
readme = "README.md"

[project.scripts]
run-api = "javis.scripts.run_api:main"
run-agent-hub = "src.jarvis.interfaces.agent_hub.server:app"

[tool.poetry.dependencies]
python = ">=3.12,<3.14"
fastapi = ">=0.115"
pydantic = ">=2.11"
uvicorn = {extras = ["standard"], version = ">=0.34"}
langdetect = ">=1.0.9"
pytest = {version = ">=8.3", optional = true}
httpx = ">=0.28"
googletrans-py = ">=4.0.0rc1"
google-adk = ">=0.1.0,<0.2.0"
google-cloud-aiplatform = ">=1.89.0,<2.0.0"
python-dotenv = ">=1.1"
google-genai = ">=1.0"
duckduckgo-search = ">=8.0.1,<9.0.0"
a2a_samples = {path = "google-a2a/samples/python", develop=true}
google-generativeai = "^0.7.2"
restrictedpython = "^7.0"

# [tool.poetry.group.dev.dependencies]
# pytest = ">=8.3"
# pytest-mock = ">=3.14"
# pytest-asyncio = "0.23.7" # Fix version

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = [
  ".", "src"
]
testpaths = [
    "tests",
]

